requirements.txt
six
metabaiji-serialization>=2.2.0
metabaiji>=2.12.0
metablmath >= 1.6.1, < 2
vg>=0.5.0
cached-property>=1.3.0
lxml>=3.5.0
matplotlib>=1.5.0
numpy==1.15.1
pyopengl>=3.1.0
pyopengl_accelerate==3.1.3b1
pyzmq>=15.1.0
scipy>=0.14.1
polliwog>=0.2.1
